Teftel Cat Stickers (animated & contrast optimized version)
Add Teftel Cat Stickers from Telegram. There are already two versions on here but one is not animated and the other has weird edges that do not look good, especially on dark backgrounds. This one has a white border for better contrast. Also it uses the correct emoji mapping like the original.